refactor and enable runtime configuration
