+++ packages/core/mesh/network-manager/package.json
@@ -41,6 +41,7 @@
     "node-datachannel": "^0.4.3",
     "p-defer": "^3.0.0",
     "simple-peer": "9.11.1",
+    "stream": "^0.0.2",
     "tiny-invariant": "^1.3.1",
     "ws": "^7.4.4",
     "xor-distance": "^2.0.0"
