+++ packages/core/mesh/network-manager/src/transport/index.ts
@@ -4,7 +4,9 @@
 
 export * from './memory-transport';
 export * from './transport';
-export * from './webrtc-transport';
-export * from './webrtc-transport-service';
-export * from './webrtc-transport-proxy';
+export * from './simplepeer-transport';
+export * from './simplepeer-transport-service';
+export * from './simplepeer-transport-proxy';
 export * from './datachannel';
+export * from './libdatachannel-transport';
+export * from './webrtc-transport';
