+++ packages/core/mesh/network-manager/src/transport/libdatachannel-transport.ts
@@ -10,7 +10,7 @@ import { log } from '@dxos/log';
 import { Signal } from '@dxos/protocols/proto/dxos/mesh/swarm';
 
 import { type PeerConnection } from './datachannel/rtc-peer-connection';
-import { Transport } from './transport';
+import { Transport, TransportFactory } from './transport';
 
 export type LibDataChannelTransportParams = {
   initiator: boolean;
